split out some more tests
